feat: Add npm publish step to release workflow
This commit adds a new step to the release-please workflow to automatically publish the package to npm when a new release is created.

The step is configured to use the NPM_TOKEN secret to authenticate with npm.

The step is conditional and only runs if the release-please action outputs that a release was created.